Backend/code/requirements.txt
aniso8601==9.0.1
boto3==1.24.11
botocore==1.27.11
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
